refactor: featuresアーキテクチャを適用
