+++ src/app/page.tsx
@@ -1,7 +1,7 @@
-import AlertContainer from '@/components/AlertContainer';
-import ForecastStatusContainer from '@/components/ForecastStatusContainer';
-import SignInForm from '@/components/SignInForm';
+import AlertContainer from '@/features/hint/AlertContainer';
 import TabBar from '@/components/TabBar';
+import SignInForm from '@/features/account/SignInForm';
+import ForecastStatusContainer from '@/features/forecast-status/ForecastStatusContainer';
 
 export default function Page() {
   return (
