+++ src/features/forecast-status/ForecastStatusContainer.tsx
@@ -1,4 +1,4 @@
-import { Pending } from './forecast-status-block/PendingStatus';
+import { Pending } from '@/features/forecast-status/PendingStatus';
 
 const ForecastStatusBlock = () => {
   return <Pending current={4} max={10} />;
