+++ src/features/forecast/PortionIndicator.tsx
