+++ src/features/hint/AlertItem.tsx
